kafka_data_ingestion/requirements.txt
# Algolia Python Client (v3.x - stable and well-documented)
algoliasearch>=3.0.0,<4.0.0

# Environment variable management
python-dotenv>=1.0.0

# Kafka client (for production mode)
kafka-python>=2.0.2

# Note: Standard library modules used:
# - json (built-in)
# - pathlib (built-in)
# - datetime (built-in)
# - logging (built-in)

# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
